requirements.txt
pyTelegramBotAPI
google-generativeai
firebase-admin
flask
gunicorn